Attempt to address #4
I am not able to reproduce the issue, but can see that there may be some cases where an ajax call to submit some sub-form doesn't cause the page to reload, so added a call to refresh the todo list to the most likely spot. Need to confirm with Lorna if this really fixes the issue. In general, there is a lack of tracking program flow, with "please wait" blockUI on a timer, rather than waiting for ajax response, and unclear when page is doing something in the background. Occasionally there's a several-second wait between the form submitting and the page refreshing, which may be the real issue.