Improve form layout for closing contracts.
Started uinder Issue #13, but more related to Issue #24. The close-contract dialog's content was formerly generated by JavaScript in site.js. This resulted in some poorly-formatted HTML and bad layout. Moving to a partial because at some point it seemed this would also be used on encumbrance page, though it isn't apparent how that gets invoked.